Move all python functions into a single file and arrange alphabetically
No functional changes
